T-shirt shop project
Implemented a project that deals with a T-shirt shop. There is an initial stock in the store.
The initial stock must contain unique items. If it is with unique elements, then there can be done
some selling-buying T-shirts. For every such transaction we know the number of pieces
traded (sold or bought). Show separate reports in different files that put in
tracking the current stock.
A T-shirt is characterized by:
- code
- color
- manufacturer
- price
- size
The original stock is read from a text file.